Setting picture size (temp fix)
Adding a quick fix to set the picture size to 960 x 720 if it's a valid size for the camera.  This is to fix pictures being too large when taken by the S3.  Need to create an issue to go back and set the size for all cameras to be something reasonable.